Refactor experience customization components to use centralized types and improve structure
