+++ packages/core/stories/patterns/experience-customization/PreferencesNavigation.tsx
@@ -1,0 +1,40 @@
+import {
+  VerticalNavigation,
+  VerticalNavigationItem,
+  VerticalNavigationItemContent,
+  VerticalNavigationItemLabel,
+  VerticalNavigationItemTrigger,
+} from "@salt-ds/core";
+
+interface PreferencesNavigationProps {
+  items: string[];
+  location: string;
+  onChange: (location: string) => void;
+}
+
+export const PreferencesNavigation = ({
+  items,
+  location,
+  onChange,
+}: PreferencesNavigationProps) => {
+  return (
+    <VerticalNavigation
+      aria-label="Experience customization sidebar"
+      appearance="bordered"
+      style={{ minWidth: "30ch" }}
+    >
+      {items.map((item) => (
+        <VerticalNavigationItem key={item} active={location === item}>
+          <VerticalNavigationItemContent>
+            <VerticalNavigationItemTrigger
+              onClick={() => onChange(item)}
+              render={<button type="button" />}
+            >
+              <VerticalNavigationItemLabel>{item}</VerticalNavigationItemLabel>
+            </VerticalNavigationItemTrigger>
+          </VerticalNavigationItemContent>
+        </VerticalNavigationItem>
+      ))}
+    </VerticalNavigation>
+  );
+};
